+++ CHANGELOG.md
@@ -6,6 +6,18 @@ The format is (loosely) based on [Keep a Changelog](http://keepachangelog.com/) 
 
 ## Unreleased
 
+## [v1.9.1] - 2025-06-16
+
+### Added
+
+- Added display of failed schema information in the validation output ([#134](https://github.com/stac-utils/stac-check/pull/134)) 
+- Added recommendation messages to guide users when validation fails ([#134](https://github.com/stac-utils/stac-check/pull/134))
+- Added disclaimer about schema-based STAC validation being an initial indicator of validity only ([#134](https://github.com/stac-utils/stac-check/pull/134))
+
+### Changed
+
+- Updated validation output to show "Passed" instead of "Valid" for accuracy ([#134](https://github.com/stac-utils/stac-check/pull/134))
+
 
 ## [v1.9.0] - 2025-06-13
 
@@ -252,7 +264,8 @@ The format is (loosely) based on [Keep a Changelog](http://keepachangelog.com/) 
 - Validation from stac-validator 2.3.0
 - Links and assets validation checks
 
-[Unreleased]: https://github.com/stac-utils/stac-check/compare/v1.9.0...main
+[Unreleased]: https://github.com/stac-utils/stac-check/compare/v1.9.1...main
+[v1.9.1]: https://github.com/stac-utils/stac-check/compare/v1.9.0...v1.9.1
 [v1.9.0]: https://github.com/stac-utils/stac-check/compare/v1.8.0...v1.9.0
 [v1.8.0]: https://github.com/stac-utils/stac-check/compare/v1.7.0...v1.8.0
 [v1.7.0]: https://github.com/stac-utils/stac-check/compare/v1.6.0...v1.7.0